Fix AddTransactionScreen account picker
A selected destination account will be cleared if user switches to an
transaction type as outgoing and a selected source account will be
cleared when user selects transaction type as incoming. Whenever type
of transaction is changed, the selected transaction category is cleared.

Additionally, the transaction category picker dropdown shows an option
to navigate to AddCategoryScreen with a pre-selected transaction type.